+++ services/scribe/api/suite_test.go
@@ -85,14 +85,14 @@ func (g *APISuite) SetupTest() {
 		}, g.metrics))
 	}()
 
-	hostName := fmt.Sprintf("127.0.0.1:%d", port)
+	hostName := fmt.Sprintf("localhost:%d", port)
 	baseURL := fmt.Sprintf("http://%s", hostName)
 
 	g.gqlClient = client.NewClient(http.DefaultClient, fmt.Sprintf("%s%s", baseURL, server.GraphqlEndpoint))
 
 	config := rest.NewConfiguration()
 	config.BasePath = baseURL
-	config.Host = baseURL
+	config.Host = hostName
 
 	g.grpcRestClient = rest.NewAPIClient(config)
 	rawGrpcClient, err := grpc.DialContext(g.GetTestContext(), hostName, grpc.WithTransportCredentials(insecure.NewCredentials()))
@@ -118,6 +118,7 @@ func (g *APISuite) SetupTest() {
 		res, realRes, err := g.grpcRestClient.ScribeServiceApi.ScribeServiceCheck(g.GetTestContext(), rest.V1HealthCheckRequest{
 			Service: "any",
 		})
+
 		if err == nil {
 			defer func() {
 				_ = realRes.Body.Close()