+++ packages/rest-api/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.6](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.0.5...@synapsecns/rest-api@1.0.6) (2023-07-10)
+
+**Note:** Version bump only for package @synapsecns/rest-api
+
+
+
+
+
 ## [1.0.5](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.0.4...@synapsecns/rest-api@1.0.5) (2023-06-13)
 
 **Note:** Version bump only for package @synapsecns/rest-api